+++ lib/index.d.ts
@@ -20,6 +20,16 @@
 // With great contributions to @akim95 on github
 
 declare namespace Realm {
+    interface CollectionChangeSet {
+        insertions: number[];
+        deletions: number[];
+        modifications: number[];
+        newModifications: number[];
+        oldModifications: number[];
+    }
+
+    type CollectionChangeCallback<T> = (collection: Collection<T>, change: CollectionChangeSet) => void;
+
     /**
      * PropertyType
      * @see { @link https://realm.io/docs/javascript/latest/api/Realm.html#~PropertyType }
@@ -124,6 +134,12 @@ declare namespace Realm {
          * @returns number
          */
         linkingObjectsCount(): number;
+
+        /**
+         * @param  {(collection:any,changes:any)=>void} callback
+         * @returns void
+         */
+        addListener(callback: CollectionChangeCallback<T>): void;
     }
 
     const Object: {
@@ -135,16 +151,6 @@ declare namespace Realm {
      * @see { @link https://realm.io/docs/javascript/latest/api/Realm.Collection.html#~SortDescriptor }
      */
     type SortDescriptor = [string] | [string, boolean];
-
-    interface CollectionChangeSet {
-        insertions: number[];
-        deletions: number[];
-        modifications: number[];
-        newModifications: number[];
-        oldModifications: number[];
-    }
-
-    type CollectionChangeCallback<T> = (collection: Collection<T>, change: CollectionChangeSet) => void;
 
     /**
      * Collection